server/portfolio/serializers.py
# portfolio/serializers.py

from rest_framework import serializers
from .models import Portfolio, Reply

class PortfolioSerializer(serializers.ModelSerializer):
    class Meta:
        model = Portfolio
        fields = '__all__'

class ReplySerializer(serializers.ModelSerializer):
    class Meta:
        model = Reply
        fields = '__all__'