server/requirements.txt
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.3
Backtesting==0.3.3
backtrader==1.9.78.123
beautifulsoup4==4.12.2
blinker==1.6.2
bokeh==3.2.2
Brotli==1.1.0
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
ConfigArgParse==1.7
contourpy==1.1.1
cycler==0.12.0
DateTime==5.2
Deprecated==1.2.14
Django==4.2.5
django-cors-headers==4.2.0
django-redis==5.4.0
django-suit==0.2.28
djangorestframework==3.14.0
finance-datareader==0.9.50
Flask==3.0.0
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
fonttools==4.43.0
frozendict==2.3.8
gevent==23.9.1
geventhttpclient==2.0.11
greenlet==3.0.0
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
locust==2.17.0
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.0
mpld3==0.5.9
msgpack==1.0.7
multipledispatch==1.0.0
multitasking==0.0.11
numpy==1.26.0
packaging==23.2
pandas==2.1.1
peewee==3.16.3
Pillow==10.0.1
plotly==5.17.0
psutil==5.9.5
pykrx==1.0.45
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
redis==5.0.1
requests==2.31.0
requests-file==1.5.1
roundrobin==0.0.4
scipy==1.11.3
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tenacity==8.2.3
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
urlopen==1.0.0
webencodings==0.5.1
Werkzeug==3.0.0
wrapt==1.15.0
xlrd==2.0.1
xyzservices==2023.7.0
yfinance==0.2.31
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
